% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calc_auc}
\alias{calc_auc}
\title{Realiza el cálculo del AUC, TPR y FPR.}
\usage{
calc_auc(dx, dy, solo_auc = F)
}
\arguments{
\item{dx}{(vector): Vector de variables a analizar.}

\item{dy}{(vector): Vector de etiquetas.}

\item{solo_auc}{(bool): Booleano para devolver solo el AUC o una lista, útil para mejorar la eficiencia de calcular_metricas.}
}
\value{
(list): $AUC, área bajo la curva ROC; $TPR, sensibilidad o true positive rate; $FPR, false positive rate.
}
\description{
Realiza el cálculo del AUC, TPR y FPR.
}
