% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calc_entropia}
\alias{calc_entropia}
\title{Realiza el cálculo de la entropía.}
\usage{
calc_entropia(x)
}
\arguments{
\item{x}{(vector): Vector de variables discretas.}
}
\value{
(numeric): Valor de la entropía.
}
\description{
Realiza el cálculo de la entropía.
}
