% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calc_var}
\alias{calc_var}
\title{Realiza el cálculo de la varianza.}
\usage{
calc_var(x)
}
\arguments{
\item{x}{(vector): Vector de variables continuas.}
}
\value{
(numeric): Valor de la varianza.
}
\description{
Realiza el cálculo de la varianza.
}
