% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{calcular_metricas}
\alias{calcular_metricas}
\title{Realiza el cálculo de métricas para los atributos de un dataset.
\itemize{
\item Varianza y AUC para las variables contínuas y entropía para las discretas
}}
\usage{
calcular_metricas(datos, clase = NaN)
}
\arguments{
\item{datos}{(data.frame): Dataset con diferentes variables.}

\item{clase}{(int | str): Columna de la variable de la clase binaria, default=NaN.}
}
\value{
(matrix): Matriz con la varianza, AUC, y Entropia correspondiente a cada columna de datos.
}
\description{
Realiza el cálculo de métricas para los atributos de un dataset.
\itemize{
\item Varianza y AUC para las variables contínuas y entropía para las discretas
}
}
\examples{
calcular_metricas(df,"Etiqueta")
}
