% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_im_pares.R
\name{cor_im}
\alias{cor_im}
\title{Cálculo de la correlación/información mutua por pares entre variables de un dataset.}
\usage{
cor_im(datos)
}
\arguments{
\item{datos}{(data.frame): Dataset con diferentes variables.}
}
\value{
(data.frame): Dataset con las correlaciones/información mutua.
}
\description{
Cálculo de la correlación/información mutua por pares entre variables de un dataset.
}
\examples{
cor_im(df)
}
