% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretize}
\alias{discretize}
\title{Realiza la discretización tanto para un solo atributo como para un dataset completo.}
\usage{
discretize(x, num.bins, algoritmo = "anchura")
}
\arguments{
\item{x}{(data.frame | list): Un DataFrame o lista que será discretizado.}

\item{num.bins}{(int): Un número natural mayor de 1, que determina el número de intervalos.}

\item{algoritmo}{(str): Algoritmo de discretización, default='anchura'.}
}
\value{
(data.frame | factor): DataFrame o vector de valores categóricos resultado de aplicar el algoritmo.
}
\description{
Realiza la discretización tanto para un solo atributo como para un dataset completo.
}
\examples{
discretize(c(1,2,3,4),2)
}
