% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escala.R
\name{estandarizacion}
\alias{estandarizacion}
\title{Estandariza el vector x, con media 0 y desviación 1.}
\usage{
estandarizacion(x)
}
\arguments{
\item{x}{(data.frame): Vector númerico.}
}
\value{
(vector): Vector estandarizado.
}
\description{
Estandariza el vector x, con media 0 y desviación 1.
}
