% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrado.R
\name{filtrado}
\alias{filtrado}
\title{Filtrado de variables en base a las métricas implementadas.}
\usage{
filtrado(
  datos,
  clase = NA,
  min_Var = -Inf,
  max_Var = Inf,
  min_AUC = 0,
  max_AUC = 1,
  min_Entropia = -Inf,
  max_Entropia = Inf
)
}
\arguments{
\item{datos}{(data.frame): Dataset de diferentes variables.}

\item{clase}{(int | str): Columna de la variable de la clase binaria, default=None.}

\item{...}{(int): Diferentes valores de umbrales para filtrar las variables.}
}
\value{
(data.frame): Dataset filtrado.
}
\description{
Filtrado de variables en base a las métricas implementadas.
}
\examples{
filtrado(df,'Etiquetas',min_Var = 15, max_Var = 90)
}
