% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escala.R
\name{normalizacion}
\alias{normalizacion}
\title{Normaliza el vector x, con valores entre 0 y 1.}
\usage{
normalizacion(x)
}
\arguments{
\item{x}{(data.frame): Vector númerico.}
}
\value{
(vector): Vector normalizado.
}
\description{
Normaliza el vector x, con valores entre 0 y 1.
}
