% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_auc}
\alias{plot_auc}
\title{Función para trazar la curva ROC.}
\usage{
plot_auc(TPR, FPR)
}
\arguments{
\item{TPR}{(vector): Vector de tasas de verdaderos positivos.}

\item{FPR}{(vector): Vector de tasas de falsos positivos.}
}
\description{
Función para trazar la curva ROC.
}
\examples{
plot_auc(auc$TPR,auc$FPR)
}
