% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cor}
\alias{plot_cor}
\title{Función para la representación de correlación/información mutua.}
\usage{
plot_cor(matriz)
}
\arguments{
\item{matriz}{(matrix): Matriz con las correlaciones e información mutua.}
}
\description{
Función para la representación de correlación/información mutua.
}
\details{
Esta función requiere de los paquetes, ggplot2 y reshape2
}
\examples{
plot_cor(cor_im(df))
}
