% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escala.R
\name{scale_x}
\alias{scale_x}
\title{Normalización y estandarización de variables tanto para un dataset como de manera individual.}
\usage{
scale_x(x, norm = F)
}
\arguments{
\item{x}{(data.frame | vector): DataFrame o vector númerico para escalar.}

\item{norm}{(logical): Booleano que determina el tipo de transformación a realizar, default=False (normaliza).}
}
\value{
(data.frame | vector): Entrada transformada.
}
\description{
Normalización y estandarización de variables tanto para un dataset como de manera individual.
}
\examples{
scale_x(c(-0.29, 2.11, -3.84, 10.69, 5.30),norm=T)
}
